+++ src/app/components/space-automation/atm-workflow-execution-preview.js
@@ -11,11 +11,13 @@ import Component from '@ember/component';
 import { computed, get, getProperties } from '@ember/object';
 import { promise } from 'ember-awesome-macros';
 import ExecutionDataFetcher from 'oneprovider-gui/utils/workflow-visualiser/execution-data-fetcher';
+import ActionsFactory from 'onedata-gui-common/utils/workflow-visualiser/actions-factory';
 import { inject as service } from '@ember/service';
 
 export default Component.extend({
   classNames: ['atm-workflow-execution-preview', 'loadable-row'],
 
+  workflowManager: service(),
   workflowActions: service(),
 
   /**
@@ -89,6 +91,26 @@ export default Component.extend({
       }
     }
   ),
+
+  /**
+   * @type {ComputedProperty<Utils.WorkflowVisualiser.ActionsFactory>}
+   */
+  actionsFactory: computed(function actionsFactory() {
+    const factory = ActionsFactory.create({ ownerSource: this });
+    factory.setRetryLaneCallback(async (lane, runNumber) =>
+      await this.get('workflowManager').retryAtmLane(
+        this.get('atmWorkflowExecutionProxy.entityId'),
+        get(lane, 'id'),
+        runNumber
+      ));
+    factory.setRerunLaneCallback(async (lane, runNumber) =>
+      await this.get('workflowManager').rerunAtmLane(
+        this.get('atmWorkflowExecutionProxy.entityId'),
+        get(lane, 'id'),
+        runNumber
+      ));
+    return factory;
+  }),
 
   /**
    * @type {ComputedProperty<Utils.Action>}