+++ src/app/locales/en/translations.js
@@ -43,6 +43,7 @@ import archiveBrowserModel from './utils/archive-browser-model';
 import selectorFilesystemBrowserModel from './utils/selector-filesystem-browser-model';
 import itemsSelectBrowserUtils from './utils/items-select-browser';
 import workflowActions from './utils/workflow-actions';
+import workflowVisualiser from './utils/workflow-visualiser';
 
 const translations = {
   components: {
@@ -90,6 +91,7 @@ const translations = {
     selectorFilesystemBrowserModel,
     itemsSelectBrowser: itemsSelectBrowserUtils,
     workflowActions,
+    workflowVisualiser,
   },
 };