+++ src/app/services/production/workflow-manager.js
@@ -13,11 +13,13 @@ import gri from 'onedata-gui-websocket-client/utils/gri';
 import { entityType as atmWorkflowSchemaEntityType } from 'oneprovider-gui/models/atm-workflow-schema';
 import { entityType as atmWorkflowExecutionEntityType } from 'oneprovider-gui/models/atm-workflow-execution';
 import { entityType as atmTaskExecutionEntityType } from 'oneprovider-gui/models/atm-task-execution';
+import { entityType as atmStoreEntityType } from 'oneprovider-gui/models/atm-store';
 import { allSettled } from 'rsvp';
 import { reads } from '@ember/object/computed';
 import { bool, and } from 'ember-awesome-macros';
 import { promiseArray } from 'onedata-gui-common/utils/ember/promise-array';
 import AllKnownAtmWorkflowSchemasProxyArray from 'oneprovider-gui/utils/workflow-manager/all-known-atm-workflow-schemas-proxy-array';
+import notImplementedReject from 'onedata-gui-common/utils/not-implemented-reject';
 
 export default Service.extend({
   store: service(),
@@ -73,6 +75,7 @@ export default Service.extend({
   /**
    * @param {String} taskId
    * @param {Boolean} [fetchOptions.reload=false]
+   * @param {Boolean} [fetchOptions.backgroundReload=false]
    * @returns {Promise<Models.AtmTaskExecution>}
    */
   async getAtmTaskExecutionById(taskId, {
@@ -87,6 +90,26 @@ export default Service.extend({
     });
     return await this.get('store')
       .findRecord('atmTaskExecution', taskGri, { reload, backgroundReload });
+  },
+
+  /**
+   * @param {String} storeId
+   * @param {Boolean} [fetchOptions.reload=false]
+   * @param {Boolean} [fetchOptions.backgroundReload=false]
+   * @returns {Promise<Models.AtmStore>}
+   */
+  async getAtmStoreById(storeId, {
+    reload = false,
+    backgroundReload = false,
+  } = {}) {
+    const storeGri = gri({
+      entityType: atmStoreEntityType,
+      entityId: storeId,
+      aspect: 'instance',
+      scope: 'private',
+    });
+    return await this.get('store')
+      .findRecord('atmStore', storeGri, { reload, backgroundReload });
   },
 
   /**
@@ -177,7 +200,7 @@ export default Service.extend({
 
     const onedataGraph = this.get('onedataGraph');
     const storeContentGri = gri({
-      entityType: 'op_atm_store',
+      entityType: atmStoreEntityType,
       entityId: storeInstanceId,
       aspect: 'content',
     });
@@ -193,6 +216,28 @@ export default Service.extend({
     });
 
     return { array: list, isLast };
+  },
+
+  /**
+   * @param {String} atmWorkflowExecutionId
+   * @param {String} atmLaneId
+   * @param {AtmLaneRunNumber} runNumber
+   * @returns {Promise}
+   */
+  async retryAtmLane( /* atmWorkflowExecutionId, atmLaneId, runNumber */ ) {
+    // TODO: VFS-8287 Implement
+    return await notImplementedReject();
+  },
+
+  /**
+   * @param {String} atmWorkflowExecutionId
+   * @param {String} atmLaneId
+   * @param {AtmLaneRunNumber} runNumber
+   * @returns {Promise}
+   */
+  async rerunAtmLane( /* atmWorkflowExecutionId, atmLaneId, runNumber */ ) {
+    // TODO: VFS-8287 Implement
+    return await notImplementedReject();
   },
 
   /**