+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-8283 Added showing multiple runs of workflow lanes
 * VFS-8547 Fixed minor bugs: reloading datatasets root, opening file info modal in share and opening hardlinks info tab
 * VFS-8482 Added dedicated page for privacy policy content
 * VFS-7512 Added current directory QoS/Dataset tags in file browser header and collapsing inherited tags in file rows